Lootr support: convert single instance of member to static method.
LockableLootTileEntity's static setLootTable method is the only point
that I can inject to modify the world while it's being generated;
hooking into the member setLootTable generally results in the world
either being null or being the server world -- and modifying the world
at the point results in concurrent modification issues.

All other instances of setLootTable are just being used to initialise a
tile and then write the information to a processor tag, which should be
automatically handled by Lootr's own processor.

Finally, this will only affect crates which Lootr supports. Thus, this
commit isn't strictly necessary, but the functionality ends up being
identical as the check in the static method is the same as what was
replaced.